+++ project/current-project.rkt
@@ -3,6 +3,7 @@
 (provide current-project)
 
 (require framework/notify
+         framework/preferences
          sauron/log)
 
 (define current-project
@@ -11,4 +12,5 @@
 
 (send current-project listen
       (λ (new-dir)
+        (preferences:set 'current-project new-dir)
         (log:info "current project is ~a" new-dir)))